backend/requirements.txt
oseti
mecab-python3
tweepy
python-dotenv
--no-binary :all: mecab-python3